backend/requirements.txt

# FastAPI and web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
websockets>=12.0

# LangGraph and AI orchestration
langgraph>=0.2.0
langgraph-checkpoint-sqlite>=1.0.0
aiosqlite>=0.19.0

# OpenAI
openai>=1.30.0

# HuggingFace
transformers>=4.38.0
sentence-transformers>=2.5.0
huggingface-hub>=0.21.0
torch>=2.2.0

# Data processing
pandas>=2.2.0
numpy>=1.26.0
openpyxl>=3.1.2
sqlalchemy>=2.0.27

# Visualization
matplotlib>=3.8.0
plotly>=5.19.0
seaborn>=0.13.0

# Utilities
python-dotenv>=1.0.1
loguru>=0.7.2
pydantic>=2.6.0

# Development
pytest>=8.0.0
pytest-asyncio>=0.23.5
black>=24.2.0
